+++ src/spyglass/common/common_ephys.py
@@ -495,7 +495,7 @@ class LFP(SpyglassMixin, dj.Imported):
                 "nwb_file_name": key["nwb_file_name"],
                 "interval_list_name": key["interval_list_name"],
                 "valid_times": lfp_valid_times,
-                "interval_type": "lfp_v0",
+                "pipeline": "lfp_v0",
             },
             replace=True,
         )
@@ -839,7 +839,7 @@ class LFPBand(SpyglassMixin, dj.Computed):
                     "nwb_file_name": key["nwb_file_name"],
                     "interval_list_name": key["interval_list_name"],
                     "valid_times": lfp_band_valid_times,
-                    "interval_type": "lfp_band",
+                    "pipeline": "lfp_band",
                 }
             )
         else: