+++ src/spyglass/spikesorting/v1/recording.py
@@ -275,7 +275,7 @@ class SpikeSortingRecording(SpyglassMixin, dj.Computed):
                 ),
                 "interval_list_name": key["recording_id"],
                 "valid_times": sort_interval_valid_times,
-                "interval_type": "spikesorting_recording_v1",
+                "pipeline": "spikesorting_recording_v1",
             }
         )
         AnalysisNwbfile().add(